% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ajnv_fun_cat.R
\name{Ajnv_fun_cat}
\alias{Ajnv_fun_cat}
\title{Precompute cluster-specific inverse matrices for random-effects updates}
\usage{
Ajnv_fun_cat(G, Z, W, D, sigma2, idx, K1)
}
\arguments{
\item{G}{Integer. Number of clusters.}

\item{Z}{Numeric matrix (N x q). Random-effects design matrix for one contrast.}

\item{W}{List of length G. Each element is the cluster-specific working weight
matrix W_j of dimension ((K-1)*n_j) x ((K-1)*n_j).}

\item{D}{Numeric matrix ((K-1)*q x (K-1)*q). Current random-effects covariance matrix.}

\item{sigma2}{Numeric scalar. Current residual variance.}

\item{idx}{List of length G. Each element contains the row indices of Z
corresponding to a given cluster.}

\item{K1}{Integer. Number of log-odds contrasts, i.e. K - 1.}
}
\value{
A list of length G, where each element is a ((K-1)*q x (K-1)*q)
matrix containing A_j^{-1} for the corresponding cluster.
}
\description{
Computes the inverse matrices A_j^{-1} for each cluster j in the
categorical GMERT, where
A_j = D^{-1} + (1 / sigma2) * Z_tilde_j' W_j Z_tilde_j.
}
\keyword{internal}
