% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_fun_cat.R
\name{b_fun_cat}
\alias{b_fun_cat}
\title{Update cluster-specific random effects using precomputed A_i^{-1}}
\usage{
b_fun_cat(G, Z, W, idx, y_t, fhat, Ainv, sigma2, K1)
}
\arguments{
\item{G}{Integer. Number of clusters.}

\item{Z}{Numeric matrix (N x q). Random-effects design matrix for one contrast.}

\item{W}{List of length G. Each element contains the cluster-specific working
weight matrix W_i of dimension ((K-1)n_i) x ((K-1)n_i).}

\item{idx}{List of length G. Each element contains the row indices of Z
corresponding to a given cluster.}

\item{y_t}{Numeric vector of length N*(K-1). Stacked pseudo-response.}

\item{fhat}{Numeric vector of length N*(K-1). Stacked fitted fixed-part predictions.}

\item{Ainv}{List of length G. Each element is a ((K-1)q x (K-1)q) matrix
representing A_i^{-1} for the corresponding cluster.}

\item{sigma2}{Numeric scalar. Current residual variance.}

\item{K1}{Integer. Number of log-odds contrasts, i.e. K - 1.}
}
\value{
A numeric matrix of dimension (G x ((K-1)q)) containing the updated
random effects for each cluster.
}
\description{
Computes the updated random effects b_i for each cluster i using
the precomputed inverse matrices A_i^{-1} for the categorical GMERT.
}
\details{
For each cluster i, this function computes
b_i = A_i^{-1} (1/sigma2) Z_tilde_i^T W_i (y_tilde_i - f_i),
where Z_tilde_i = I_(K-1) \otimes Z_i.
}
\keyword{internal}
